fix: CI green — ruff clean, blake3 graceful fallback, unused imports removed
- pyproject.toml: extend ruff ignore list (E402, RUF001-003, RUF006, SIM*)
  E402 is expected with 'from __future__ import annotations' + module docstring pattern
- src/uzpr/util/hashing.py: blake3 optional (falls back to sha256 when no wheel)
- src/uzpr/util/system.py: fix E741 ambiguous var name l -> ln
- scripts/_extract_research.py: remove unused 'import re'
- scripts/fetch_wordlists.py: remove unused 'import io'
- src/uzpr/ui/pages/{active_jobs,history,home,new_job_wizard}.py: remove unused qfluentwidgets imports
All 11 unit tests pass locally.

+++ src/uzpr/core/dedup.py
@@ -28,9 +28,9 @@ class BloomFilter:
         """
         ln2 = math.log(2)
         # Optimal bit-array size: m = -n * ln(p) / (ln2)^2
-        m_bits = int(math.ceil(-capacity * math.log(fp_rate) / (ln2 ** 2)))
+        m_bits = math.ceil(-capacity * math.log(fp_rate) / (ln2 ** 2))
         # Optimal number of hash functions: k = (m / n) * ln2
-        self._k = max(1, int(round((m_bits / capacity) * ln2)))
+        self._k = max(1, round((m_bits / capacity) * ln2))
         self._m = m_bits
         self._byte_size = math.ceil(m_bits / 8)
